third_party/blink/web_tests/virtual/partitioned-popins-disabled/external/wpt/partitioned-popins/partitioned-popins.cookies-allowed.tentative.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify Partitioned Popins cookie access when third-party cookie access allowed
  assert_equals: expected "ReadOnLoad:FirstPartyNone-ThirdPartyNone-,ReadOnFetch:FirstPartyNone-ThirdPartyNone-FirstPartyNonePopin-ThirdPartyNonePopin-,ReadOnDocument:FirstPartyNone-ThirdPartyNone-FirstPartyNonePopin-ThirdPartyNonePopin-,ReadOnFetchAfterRSA:FirstPartyNone-ThirdPartyNone-FirstPartyNonePopin-ThirdPartyNonePopin-FirstPartyNonePopinAfterRSA-ThirdPartyNonePopinAfterRSA-,ReadOnDocumentAfterRSA:FirstPartyNone-ThirdPartyNone-FirstPartyNonePopin-ThirdPartyNonePopin-FirstPartyNonePopinAfterRSA-ThirdPartyNonePopinAfterRSA-,ReadInPopinIframe:FirstPartyNone-ThirdPartyNone-,ReadInPopinIframeAfterRSA:FirstPartyNone-ThirdPartyNone-FirstPartyNoneAfterRSA-ThirdPartyNoneAfterRSA-" but got "ReadOnLoad:FirstPartyLax-FirstPartyNone-,ReadOnFetch:FirstPartyStrict-FirstPartyLax-FirstPartyNone-FirstPartyStrictPopin-FirstPartyLaxPopin-FirstPartyNonePopin-ThirdPartyStrictPopin-ThirdPartyLaxPopin-ThirdPartyNonePopin-,ReadOnDocument:FirstPartyStrict-FirstPartyLax-FirstPartyNone-FirstPartyStrictPopin-FirstPartyLaxPopin-FirstPartyNonePopin-ThirdPartyStrictPopin-ThirdPartyLaxPopin-ThirdPartyNonePopin-,ReadOnFetchAfterRSA:FirstPartyStrict-FirstPartyLax-FirstPartyNone-FirstPartyStrictPopin-FirstPartyLaxPopin-FirstPartyNonePopin-ThirdPartyStrictPopin-ThirdPartyLaxPopin-ThirdPartyNonePopin-FirstPartyStrictPopinAfterRSA-FirstPartyLaxPopinAfterRSA-FirstPartyNonePopinAfterRSA-ThirdPartyStrictPopinAfterRSA-ThirdPartyLaxPopinAfterRSA-ThirdPartyNonePopinAfterRSA-,ReadOnDocumentAfterRSA:FirstPartyStrict-FirstPartyLax-FirstPartyNone-FirstPartyStrictPopin-FirstPartyLaxPopin-FirstPartyNonePopin-ThirdPartyStrictPopin-ThirdPartyLaxPopin-ThirdPartyNonePopin-FirstPartyStrictPopinAfterRSA-FirstPartyLaxPopinAfterRSA-FirstPartyNonePopinAfterRSA-ThirdPartyStrictPopinAfterRSA-ThirdPartyLaxPopinAfterRSA-ThirdPartyNonePopinAfterRSA-,ReadInPopinIframe:FirstPartyNone-ThirdPartyNone-,ReadInPopinIframeAfterRSA:FirstPartyNone-ThirdPartyNone-FirstPartyNoneAfterRSA-ThirdPartyNoneAfterRSA-"
Harness: the test ran to completion.
